Changed the Camera's projection matrix to a field that gets updated when necessary to improve performance
Added outlines and inlines and a OutlineParameters helper class to accomodate them.

Moved a lot of UIObject user interactivity handling to GameObject instead.

Improved SetTextString function in the Text object by changing it to reuse as many Letters as it can when updating the string.

Moved the Particles class into the Objects folder

Added a static Colors class to make it less memory intensive to apply colors to large swaths of objects

Added the ability to resize UIBlock objects. Added an inline to UIBlock objects. Added the fog texture as a proof of concept background texture for UIBlocks

Added support for rendering to a generic frame buffer in the renderer.

Added support for MultiTexturing in objects. The second texture will be mixed onto the first texture based on a passed percentage.

Moved the Renderer class to a new Rendering folder. Added a new RenderObjectListInstancedGeneric function to reduce draw calls when attempting to render many lists of similar objects

Added frustum culling for objects outside of view of the camera. In that same vein, added TileChunks to allow large swaths of tiles to be culled without requiring each tile to be individually checked.

Added a SceneController class that loads and unloads Scenes. This also contains a scene messaging setup that allows scenes to communicate certain information and make requests with other scenes.

Fixed an issue where repeating a property animation would cause ticks to be missed.

Added a "ticks per second" counter that can be toggled on or off to property animations so that performance can be tracked.

Added functionality to cull objects every time the camera's position is updated.

Added some support for loading a bitmap as a texture.

Added a texture name to Texture object definitions. This is primarily being used to determine whether to load a secondary texture for mixing in the renderer.

Added a static "UsedTextures" dictionary to the Texture class to track which textures are currently loaded by OpenGL

Added a sample property animation to the movement ability.

Moved the BaseTile and TileMap class into a new Tiles folder

Added selection tiles and a hovered tile to TileMap. Also added tile chunks and object culling

Added fog and explored tile stuff for the tilemap in the MenuScene

Added message parsing to scenes. The messages they can parse currently are: requests to stop rendering, requests to intercept clicks, and request to intercept keystrokes. These are then tracked by flags in the Scene class and handled accordingly

Fixed clicks being applied to multiple objects in the same space

Deleted some old resources that are no longer necessary

Added multi texture and outlining functionality to the fastDefaultShader shader.

Added a set of simpleShaders intended to render the frame buffer texture to a screen sized quad

Changed the rendering process to queue up everything that will be rendered before rendering them all at once.

Moved camera movement functionality from the Window.cs file to the MenuScene

Moved texture loading from the Window.cs file to the SceneController

Changed the TickAllObjects function to wait for all tasks to complete before continuing.

Added a call to resize all in-use FrameBufferObjects when the screen is resized. The FBOs will be scaled based on the difference between the last screen size and the current so they will maintain their current ratio.

------------------------------
Frame buffer objects and frustum culling probably took about 2 years off my life to implement :)