Changed most (if not all) cases where a Vector2 was used for scaling UI component or getting the dimensions of UI components to the UIScale and UIDimensions classes
Added an input UI component that functions as an input field normally would

Added automatic rescaling to the UIList component

Added a ScissorData class to the GameObject class. This allows for regions of the screen to be selectively constrained for the items inside of them. (Particularly useful for UI components such as scrollable areas)

Moved several helper classes into their own files

Added support for scissoring in the RenderObjectsInstancedGeneric function

Added support for focusing UI components

Added "anchoring" to UI components which gives the ability to get or set a position via an anchor point instead of the center of the object. The supported anchor points are: top center, top right, top left, center left, center, center right, bottom left, bottom center, and bottom right