Added line and column limits to the input component
Added scrollable areas. These don't render any children of their BaseComponent that are outside of the bounds of the VisibleArea component. These are intended to be nested inside of other components to accommodate scrolling/large UI elements that take up a significant portion of the screen

Added the ability for focused UI components to gather the mouse state when the frame is updated

Created a ScrollableArea demo