Renamed several ambigious/poorly named fields and methods
Changed the SceneID to use the same ID method as GameObjects

Added a couple more Scene event handlers for object focusing and, if needed, when rendering completes

Added support for energy costs to the Move ability. As it stands right now each ability will need to manage energy costs and expenditure themselves.

Added a UnitStatusBar UI component that displays the name of a unit over their head.