Added a TimedHover event that triggers after an object has been hovered for 500ms
Started adding lists of generic event actions that can be added to objects on demand. More of these will be added as the need arises

Added a check to UIObject.BoundsCheck to prevent any operations using the ReverseTree from continuing if the tree has been changed during the loop.

Changed the Renderer class to static. Added a way for GameObjects to load any textures for objects they create themselves.

Converted the Footer to a GameFooter. Added some unit information to the footer (name, health, and shields) along with tooltips. Buffs (once implemented) and available abilities will be added to this list soon.

Added a skeleton unit

Added some further handling for unit damage dealing and kills