Added a spiderweb icon
Revamped the ScrollableArea component to actually work.

Changed how scissored objects are rendered to render the first block of the object list, then render the scissored object, then render the second block of the object list and so forth. This fixes an issue where scissored objects were being rendered first and would, as a consequence, always be in front of other objects.

Added a new ContextManager class to handle flags in a cleaner way than just having a billion random public booleans lying around.

Moved tooltip creation in the CombatScene and added a transparent UIBlock with a high ZIndex specifically to hold tooltips

Added an EventAction dictionary to hold actions for hotkeys. Might need to expand this to hold multiple actions per EventAction but for now one each will suffice.

Moved some ability handling from the Move and Strike abilities to the main Ability class.

Added Buffs and buff handling to Units

Added a slow ability that utilizes the new Buff class

Added a display for Buffs in the GameFooter. This display can handle any arbitrary amount of buffs.

Made the UnitStatusBar component a bit smaller overall