Added a cleaner way to create the ability icon hover effect and put it in the UIHelpers class. I might change the location at some point but UIHelpers will suffice for now.
 Changed the EnergyDisplayBar to update the maximum energy it displays per unit.

Temporraily removed the SelectedTiles set up for determining where an ability can be cast. It doesn't look very good and there is definitely a better way to accomplish it.

Added a UnitSelectionTile object that sits below a unit to display if that unit is targeted or selected. This way multiple units can be targeted/selected at the same time and easily managed.

Added an energy boost modifier to Buffs. These are used to increase or decrease a unit's energy directly.

Added a neutral health bar color

Changed the outline handling in the basic fragment shader to use distance fields. This will most likely be expanded upon in the future.