Added a new TileSpritesheet resource that will contain all of the tile textures that I'll be using for the moment.
Copied a DirectBitmap implementation from stack overflow. Using this I can store the TileSpritesheet texture into static memory for faster reading.

Loaded the TileSpritesheet into a static variable in the TileMapController.

Added a TileTexturer class that uses that static bitmap and a passed tilemap to create a texture version of the tilemap. This comes with some insane performance gains. Namely a 100x100 tilemap done the old way was getting ~150fps whereas this new implementation gets ~4000. These changes still need to be generalized and added in to the render pipeline.