Renamed DirectBitmap to StaticBitmap to indicate more clearly that the bitmap is being stored in static memory
Changed the tilemap render process to render a generated texture onto a quad instead of rendering every tile every frame. This uses a PixelBufferObject to store the texture on the graphics card and then it gets updated whenever a tile needs to be changed. Currently there is a slight lag associated with this and I am researching a solution (namely making better use of TexSubImage2D) (might also want to create a dedicated shader setup to do this via opengl instead of using the CPU to map the pixels)