Removed unused particle shader.
Reworked the tilemap dynamic texturing system to use the GPU instead of the CPU. This process uses a newly created shader pair (tileMapShader) to render the tiles in the correct locations. In this process a framebuffer is created for each tilemap and whenever a tile has changed only the changed tiles will be rerendered onto the framebuffer. This texture is then rendered onto a texture quad in the same way as before.
Note: This way of doing things is exponentially faster than updating the information on the CPU and it also stores the textures on the GPU so there isn't a huge hit to memory. I'm pretty proud of this one.