Fixed an instance where TextComponent anchor positioning was not working correctly
Added some more functionality to the CreateTooltip UIHelpers method and created a tooltip that appears when hovering over a tile.

Fixed the issue causing the tile map texture to go crazy when resizing the screen. Now when the screen is resized the texture is reinitialized.

Added a Dispose method to the FrameBufferObject class. This will free up the GPU memory holding the texture, depth buffer, and framebuffer.

Added a global Settings class to handle general purpose flags that won't need to be saved.

Updated a couple functions to use custom structs for parameters due to parameter bloat

Added height to tiles and consideration for height in pathing and vision. Added a HeightIndicatorTile that displays when a tile is out of vision due to height differences

Fixed the EscapeMenuScene opaque backdrop issue by sending it to the renderer as a generic object instead of a UI object. Rendering it as a UI object was causing it to render first and therefore obscure everything behind it.

Created a place to put custom tile maps so that new features can be tested outside of the base TileMap class.

Moved a handful of BaseTile fields into a new TileProperties class.

Added a static FeatureGenerator class that will be used for testing feature generation for the moment.

Separated outlined tiles and non outlined tiles when updating the tilemap framebuffer. The non outlined tiles are rendered first and then the outlined tiles second. This way the outlines stay crisp when the tile is bordering an object without an outline.