Add an interface that denotes which objects are tickable. This is used to create a list of tickable objects for scenes that don't require  them to be in the render pipeline.
Added a new constructor for GameObjects that allows for a single frame object to be easily created

Removed the RenderObjectListInstancedGeneric function and will now instead be calling RenderObjectsInstancedGeneric on each list instead. At the moment it is more performant for structures and units which are the primary concerns. If we eventually need to display a significantly larger amount of objects this might need to be reworked.

Moved all render queue functions and lists into a new static RenderingQueue class because the Renderer class was getting fairly bloated

Fixed a handful of resizing related issues.

Added structures and cliffs and fleshed out tile heights effect on vision and abilities. Structures are a derived class of unit and will generally fill the role of anything mostly static that is attached to a particular tile. In this case I am using it for trees and cliffs.

Changed the map to be explored by default.

Added some structure information to the tile tooltip.

Change the TileChunk dimensions for TileMaps from 10x10 to 5x5. After some not-so-rigorous testing it seemed to perform better for large numbers of structures/units.

Added height information to units and structures. This information is used to determine whether the object will obscure vision of the tiles beyond it.

Added a new spritesheet for structures