Added support for right click menus to GameObjects
Added a queue for "low priority" game objects. In this case it's being used for the escape menu's background modal but I might come up with a better way to handle this in the future.

Added an optional parameter to update all tiles on the dynamic texture when calculating fog.

Added the concept of context managers for managing event states. The usage of it here is for storing whether a context menu is open when handling a click. This is useful because it allows us to make decisions based on the initial state instead of the current state when handling events.

Changed all instances of onEvent to OnEvent for clarity.

Added a priority field to scenes. This determines which scene will have events applied to first.

Cleaned up a lot of compiler warnings.

Introduced the concept of "pathable height" when dealing with tiles. If a tile's structure is pathable then it will take the structure's height into account when calculating whether it can be pathed through and otherwise it will just take the tile's elevation.

Added a FeatureEquation class and an example of FeatureEquations being used to generate features throughout several tilemaps. FeatureEquations should contain the tile maps that they will affect and any information required to maintain continuity with adjacent tile maps.

Added a function in TestTileMap for generating contiguous walls. This should be relocated to a new home at some point.

Added an example of a right click menu to BaseTiles. At the moment this uses the Tooltip object but this could potentially require a custom class eventually.

Changed the height based color correction to calculate it's values based on a tile map's "BaseElevation" (this BaseElevation value will need to be managed for every tile map, most likely on feature generation)