Added better selection handling for the Button ui component.
Created a method in UIHelpers to simplify the process of creating a simple right click context menu.

Added a new menu that opens when pressing Tab. This menu functions similarly to tooltips on the back end but is larger and contains multiple tabs.

Added enum extension functions to simplify getting ui strings for a couple enums.

Changed the 256x256 texture to 128x128 for real this time (kept the same overall size of the spritesheet so instead of 100 tile sprites we can store 400).

Actually fixed the issue with wonky display textures. The culprit ended up being in the tile map vertex shader where the clip space tile sizes were being calculated. I was incorrectly using the actual dimensions of the tile instead of the dimensions of the spritesheet cell and because of this the 128x128 texture was being mapped onto something like a 124x92 space which was not so good looking.
As a side effect to that texture fix, a lot of unnecessary usages of the aspect ratio have been fixed. It'd essentially allocate extra space on the framebuffer texture to account for aspect ratio, use up that space for no reason in the shader, and then scale by the inverse of the aspect ratio when creating the quad. No idea what I was thinking with that one