Moved cube related functions into their own static class
Added several new icons into the IconSheetIcons enum

Added a setup for adding property animations that works in the same way as the property animation destruction queue.

Moved most game related Unit properties into a new UnitInfo class

Added a PostTickAction to Scenes that gets invoked after all objects are ticked. This is useful as a method of queuing actions that might change tickable objects.

Added a Shoot ability that acts as a physical ranged attack.

Added a handful of helper functions into various abilities to help with AI usage of the abilities.

Added a forest feature equation that creates trees on grass tiles in a certain radius

Changed how feature equations work to now hold the tiles that they've determined are useful in a hashmap. This allows for O(1) seeking of a potentially very large amount of tiles when loading new tilemaps.

Changed the river feature equation to use this new feature equation setup.

Added a way to load surrounding tile maps and apply loaded feature maps to them. This allows the player to seamlessly (with a small loading time) walk around a persistent world.

Changed pathfinding to use a version of the A* pathfinding algorithm. This is substantially faster than the previous method I was using and also allows for individual tiles to have different energy costs to traverse.

Added a UnitAI class to Units that determines how their combat turn is taken.

AI controlled units use sets of "Dispositions" to determine an action that they will take (with an action being the usage of one of their abilities). These Dispositions create actions based on factors such as enemy position, enemy health, allied health, etc and apply a weight to each action based on the overall weight of their disposition. The basic action that every AI considers first is ending their turn and this holds a weight of 1. So, if an action has a weight higher than 1 then it will be taken, otherwise their turn ends.