Changed the Cube coordinate midpoint rounding from "AwayFromZero" to "ToEven". I think it looks better.
Added a method to disable a Button component

Created a SetCameraPerspective method for Icons

Added functionality for starting and stopping combat. Being out of combat removes the energy cost of abilities and also changes a couple parts of the UI. Additionally the player's vision is increased outside of combat to aid exploration.

Changed fog to be set for each team independently in the same way that explored tiles are saved. This is needed for AI calculations.

Added a new QueuedObjectList<T> object that mirrors the functionality of a List<T> but contains 2 internal queues for adding and removing objects from the list. These queues are then evaluated during the render cycle when necessary. This is a preferable way of doing things since ant object changes can now be done asynchronously with no worries about changing a list during enumeration. An AddImmediate method does exist, however, for directly adding an item to the list (in case that is ever necessary).

Added a method in the SceneController to initiate an object culling call.

Added functionality to determine whether an ability is in ranged from a specific point as opposed to the CastingUnit's position.

Moved energy calculations into the OnCast methods of abilities.

Added an animated effect when an ability is cast that makes it easier to tell when the AI is doing things.

Added a 750ms delay to the AI's actions for clarity.

Added some dev buttons to the TabMenu component

Changed the TilesToUpdate queue to be async friendly by creating an alternating set of 2 queues. At any given time 1 queue gets filled with the tiles to be updated and when the dynamic texture gets updated the selected queue gets swapped in while the other gets evaluated. This allows tiles to be added at any time with no worry of causing an issue with the render context. (this also allows things like vision and an AI's action to be calculated without blocking the render thread which is very nice)

Changed the GetVisionInRadius and similar methods to use a HashSet instead of a List to remove the possibility of duplicated values.

Added a new AIAction, MoveInRangeOfAbility. If not in a valid space to use the passed ability it will find a valid space or move closer if none is found. Implemented this change for the RangedDamageDealer disposition.

Changed unit vision radius into a getter. Eventually I'll work buffs into this calculation as well.

Moved AI action calculations into a new task to increase performance. (this mage the QueuedObjectList changes necessary)

Added an EnableTileMapUpdate context flag to scenes to allow multiple related tile changes to be completed and batched together before rerendering the texture. In this case it's used for fog calculations to avoid a split second where the full fog layer is updated before the team fog is filled in.