Changed the UIObject Children tick implementation from an enumerator to a for loop so that the list can be changed on demand and it'll just fail gracefully.
Changed how a scene's InitiativeOrder was stored from a list to a queued list.

Added a TemporaryVision class that is used to give vision over an area for a limited duration. This duration can be ticked down in a couple different ways depending on what is needed.

Implemented a stealth/hiding feature. A stealthed unit will not be visible to other teams if the difference between the unit's stealth skill is significantly greater than the other team's units' scout skill. View the stealth implementation issue for more details.

Added an OnRender method to Scenes to simplify the updating of queued lists.

Changed the QueuedObjectList to inherit a QueuedList class that overrides the Add method to load new textures.

Added icons to go with the stealth feature

Changed the Move ability to not break stealth

Added a stealth buff to display when a unit is in stealth. This buff has no other functionality

Added a context manager to abilities. Currently this is being used to conditionally void the energy cost of the ability or to skip the hovering icon animation. This will be used to store short term context flags that help determine how an ability should function in non-standard cases

Fixed a couple bugs in Ability.TrimTiles regarding unit targeting

Changed the UnitStatusBar.SetWillDisplay method to only update the unit status bars when the display value has changed. This appears to have fixed the random giant shield bug.

Changed the RangedDamageDealer to weight attacking higher than moving in range.

Added a Hidden and Scouting class to UnitInfo. These hold the skill levels and several helper functions related to scouting and hiding.