Changed RenderableObjects to use a BaseColor and any number of AppliedColors that then get combined into a single InterpolatedColor for rendering. This is easier to keep track of for temporary modifications to an object's color although not many areas use this functionality yet.
Added a Color class that holds basic color information as well as a boolean to determine whether this color should be applied or not. The reason I'm using this custom class instead of a Vector4 or a System.Drawing.Color is because those are both structs and can't be manipulated in the same way that classes can (as far as I know anyways). With this setup, anything giving an object an applied color can keep a reference to the color and easily change/remove it without a lot of hassle.

Added a try/catch to the UIObject.BoundsCheck method. This method hasn't had any issues for a while but I did run into a random crash because of it and because of that it is probably better to handle fails silently here instead of killing the program.

Fixed am off-by-1 error in UIObject.RemoveChildren that was causing the first Child object to not be removed.

Moved UI objects from first in the rendering queue to last (but after a depth buffer clear). This makes transparent/semi transparent UI elements possible since there will actually be something rendered before them to display.

Added a simple TurnDisplay component that displays all the units that are currently in combat and when they will take their turn.

Changed the KeyboardState and MouseState fields in the Scene class to use getters instead of having to be manually set every frame. This way of doing things is waaaay easier to manage.

Added Play and Pause functionality to the Animation class. Not really sure how I didn't add this on the first pass.

Changed the QueuedList to use a for loop instead of a foreach enumerator. I should potentially change this to use a "double barrel" queue approach for adding and removing items similar to what was done for updating tiles.

Changed the Hide ability to use an AppliedColor instead of altering the base color of the object.

Changed the EnergyDisplayBar to use an AppliedColor on the color shifting property animation.

Added some async safety to updating the footer with unit information. If something attempts to update the info while it is currently being updated then that action will be queued for when the current update finishes.

Added an option in the escape menu to toggle vsync

Added a CreateBaseObject method to Units so that copies of the unit can be created on-demand. There is most likely a better way to do this but for the moment this will suffice.

Moved the cursor render call to the end of the render cycle and added a depth buffer clear before it.