Added a flag in GameObjects to determine whether they have a context menu associated with them. This will need to be expanded on again when context menus could potentially be state dependent (ie whether a unit is dead or not)
Fixed an issue where a tooltip/context menu would be placed off of the screen when the mouse was too close to the top or right edges of the screen. Now it will change the anchor position of the tooltip in those circumstances.

Added some nicer handling for ability usage during turns, unit selection, and footer UI. Now abilities can not be used while another ability is in progress and, similarly, abilities cannot be used on a unit if they are not the scene's current unit.

Added a "Venture Forth" button to load the surrounding tilemaps. This button will appear when the main unit is within 5 tiles of the edge of the loaded tilemaps.

Made tilemap selection tiles into a QueuedList. Also removed selection tiles from the object culling process.

Removed right clicking on units. Right clicks will be handled based on the tile that was right clicked instead of trying to figure out whether a structure/unit/tile was supposed to be clicked.

Added a Finish method to property animations. This stops the animation and invokes the OnFinish action.

Changed the QueuedList to coopt the "double barreled" queue approach that was used for TilesToUpdate. This should mean that all add and remove operations done to a queued list will be preserved.

Added support for stopping a move action at any point of the path. Also changed the way energy is spent for the move action to update on a per-tile basis instead of all at once.

Added tooltips to abilities although none have been implemented yet.

Added a path feature. The path works similarly to the river in that is contains a start point and any number of stops. Instead of water the path uses stone tiles for the main path and wood tiles for any bridges over water.

Added a AddMeanderingPoints method to the PathParams struct. This creates decent looking wavy patterns in a path or river. To further expand on this the meander cycles could be made longer to increase the wavyness of the feature.

Added a UnitOnTile field to BaseTiles to track the unit that is currently on the tile. Along with the HasContextMenu change this is used to display a list containing the various context menus that are available when right clicking a tile. Selecting an option from that list then shows that option's context menu.

Added Move ability path calculations to a task. This hides the long calculation times when attempting to path to an obstructed area. Added a bit of wiggle room to the path calculation when looking for a path. It will now consider paths 1.5 times longer than what the energy cost would suggest due to the possibility of encountering tiles that decrease the energy cost to move through.

Made it so that a unit's move action will be canceled if that move reveals a hidden unit.