This is a large one.
Changed all instances of BaseObjects.Add to a new virtual AddBaseObject method that allows for classes to implement their own functionality when adding objects.

Removed outlining from the instanced shader. Might revisit this at some point.

Added a render queue for lighting quads. These are rendered after everything except UI. The depth buffer is cleared as well so the color is in front of everything.

Added a day night cycle. This updates a static EnvironmentColor variable in a constant interval, transitioning between 4 time of day colors. This environment color is then used as the base color for the lighting process.

Added 4 backdrops (I called them modals for some reason) that surround the tilemap cluster and sit flush with the lighting quad. These backdrops contain the current environment color and extend farther than the player will be allowed to move their camera. Their purpose is to apply the environment color farther without requiring a larger lighting texture.

Added lighting. This encompasses 2 new sets of shaders and several new classes. The first set of shaders and class represents light obstructions. These use a light obstruction spritesheet to help the lighting shader determine what to do with the light beam when encountering them. At the moment only a single mode is supported; immediately kill the beam when an obstruction is impacted.

The next set of shaders takes light generators and determines the area that they light up using the saved obstructor data. These light generators are then rendered to an FBO with the environment color as the backdrop (and some special alpha blend functions) then thrown onto a quad lined up with the default 3x3 tilemap cluster setup.

Added functionality for something I call "unit collating". What this does is gather all structures and units into a single render list so that they can easily be passed to the renderer when necessary. This list is automatically updated when structures/units are created and destroyed.

Added OnChange functionality to the custom Color class. This allows different parts of the game to hook into, for example, the EnvironmentColor and update something accordingly.

Changed the default Nearest filtering option from NearestMipmapNearest to NearestMipmapLinear. This fixes an issue where the mipmaps would look "shiny" when at certain angles/distances.

Added a "turbo" dev setting for movements. This allows the user to teleport when using the movement spell.

Changed UnitTeams to use a Relation system to determine whether to be friendly, neutral, or hostile instead of baking that into the UnitTeam itself. This will allow for much more nuanced handling where there could be many teams all with different relations to each other fighting instead of just good guy vs bad guy.

Added a graveyard feature. This uses a new idea where the affected points value stores a bit array instead of just an enum value. This allows for many different generated features to exist on a single feature point.

Added a grave structure to go along with the graveyard feature.

Added object clean up handling to BaseTiles. This will remove the structure and unit attached to the tile (if applicable) and invoke their clean up handling.

When clearing a chunk, each tile will now have their clean up handling invoked before emptying the chunk.

Moved from 5x5 tile chunks to 10x10 tile chunks. This resulted in a pretty significant gain in FPS (something like 200 -> 350 for the worst case scenario). I do not remember if that was before or after the unit collating feature however.

Made tile map cluster generation "circular". When a new set of tile maps are created each tile map will be moved so that the positioning of the maps always stays the same. Each unit and structure is moved as well as the camera.

Added a new UnitSearchParams class for UnitAI handling. This allows for parameters regarding the unit to be programmatically queried. If the unit matches all queries then true is returned, false otherwise.

Added a SimplifiedTileType enum. This allows for all of the sub categories of a tile (such as grass, grass_2, altgrass, etc) to all be identified as being grass.

Moved all old shaders into an OldShaders folder. This is just useful to separate currently relevant shaders since the folder is getting pretty cramped.