Revamped vision to use a ray marching algorithm similar to lighting. The VisionMap creates its own version of the obstructions map but with a resolution of 2x2 instead of 32x32 for each object (the resolution doesn't actually matter too much since all we need is a very rough approximation to make it convincing). The VisionMap static object updates every time a unit or structure has its tile map position changed.
Created new VisionMap methods to replace the old vision methods and updated FillInTeamFog to use the VisionMap.

Fixed the issue with being unable to select a controlled unit's ability if they weren't on the player unit's team.

Added an ActionQueue class that simplifies the asynchronous queued action function pattern I've been using. When you begin the action you StartUpdate and when you end it you EndUpdate. Before starting the update you check whether an update is in progress and if one is then you either AddAction (for cases where there are parameters that need to be acted on) or AddActionSingle (for cases where the method has no parameters and doesn't benefit from being called multiple times in a row).

Fixed some lag being caused by selecting the Move ability by removing an unneeded GetValidTileTargets call.

Changed the Shoot.UnitInRange method to use the VisionMap vision functionaltiy instead of the old vision functionality.

Added a TrimUnits method to Abilities. This functions in pretty much the same way as TrimTiles but is geared towards abilities that can't target tiles and therefore doesn't benefit from getting all valid tiles.

Made VisionGenerators and LightObstructions intrinsic to units (and, by extension, structures). If a structure or unit shouldn't block vision then their LightObstruction should have type None. If a unit or structure shouldn't generate vision then their VisionGenerator should have a radius of 0 or a team of Unknown.

Added some functionality in TileMapController for dealing with a tile in terms of its cluster position (its position on the 3x3 map)

Tweaked the lightShader to generate softer shadows. When a light obstruction is met the light will now decay significantly faster instead of immediately dropping to 0. This makes the lighting look a lot more natural and less jarring.