Added the NVorbis dependency for decoding ogg files.
Added support for playing audio. To play audio, sound files need to be loaded into AudioBuffers and then passed to the Sound class. The Sound class will communicate with the SoundPlayer static class and be allocated a source.
Once this source has been assigned to a Sound that Sound must either play or specifically state that it is keeping the source.
If an active source is not playing any audio and isn't being kept alive by its Sound then it will be deallocated and returned to the source pool.

Added a TileMapLoadInProgress flag to scenes so that the scene will not attempt to apply fog to the map before all maps are loaded and all vision can be calculated.

 Added an EffectEnded call to the Hide ability.

Created what should be a viable solution for the FeatureEquation.AffectsMap method for the time being. This setup simply loops through each tile in the map and checks in an affected point exists. This might become prohibitively slow once many features need to be accounted for but that can be optimized later.

Moved unit click handling into the tile click handling. This way selecting a unit will always be predictable and standardized.

Fixed a memory leak caused by units not being decollated when destroyed.

Added a copy step to the VisionMap.CalculateVision method. This copies all the vision generators over to a new list before calculating the vision which fixes issues caused by the list being modified while vision is in the process of being calculated.

Made it so that a unit will always have vision over something that is directly blocking their vision.