Added support for automatically loading and unloading sound buffers. This is controlled by a watchdog that checks every loaded buffer every 30 seconds and unloads it if no active sounds are using it.
Added a handful of new properties to control sounds.

Changed all instances where I was using lists of actions to now use the "event" feature of C#. Honestly I'm not really sure why I didn't look into that to begin with.

In areas where necessary, locks on objects have been added to increase thread safety. Locks more or less solve most of the thread safety issues I was encountering.

Made a handful of changes to increase general performance.

Added an Extensions class to the Engine_Classes folder. This will be where class extensions for default classes go. (for example, the first extension was to add new functionality to the List class)

Added a "change token" to queued lists. This increments every time the list is changed and is intended to be a quick approximation to cut down on more  expensive checks over the whole list.

Added ability tooltip functionality.

Added sound effects to a lot of things.

Added combo abilities.

Added hotkey indicators to ability icons.

Began the process of adding a scrollable area for abilities similar to the buff area. Some issues are still being worked out with that though.

Added a "flag" messaging system to the scene messager. This allows a specific enum value to be sent as a message which should hopefully make the system more useful.

Added a Sounds static class that contains the paths to AudioBuffers containing sound effects.

Made a couple performance improvements to the VisionMap methods. These are in a good spot for the moment.

Added a boatload of sound effects.

Moved the SoundPlayer initialization into the Main method and initialized it on it's own thread. This appears to have solved the crackling audio problem.

Added a GlobalRandom object. This is an object that be used for non-consequential random number generation like which sound to play when stepping on grass and etc,

Changed the tick timer to not longer be dependent on the OnRender function. It generally wasn't a problem but consistently higher or lower framerates would cause noticeable differences in tick rates due to polling frequency. IE a high FPS would poll faster and therefore see that it passed the tick threshold faster. It wasn't a huge issue since ticks per second would be consistent either way but fixing it is just better for future-proofing