Fixed BaseObjects not being placed relative to their GameObject correctly.
Changed the GameObject LoadTexture method to always load the textures during the render thread to avoid async errors.

Updated particle generators to have an OnFinish function as well as a way to specify whether the generator should repeat or not.

Created a shader specifically for rendering particles. This uses a lot less of the extra functionality needed for general rendering so it doesn't hurt performance to have many particles.

Added OnRender and OnRenderEnd events to the static Renderer class that can be subscribed to. When using these it's very important that the actions are unsubscribed asap.

Added a general purpose _particleGenerators list to Scenes. Usually a particle generator will be tied to a GameObject but there could exist a case where a particle generator would need to exist separately.

Added a visual effect to the Shoot ability.

Added a new DamageInstance class for usage with abilities and buffs. This is used to package all information pertinent to a single instance of damage into one place. And similar to things like Dota a single damage instance can contain multiple types of damage.

Moved the ability's EffectEnded method into a task. This is required because calling the EffectEnded method from something like a property animation that gets called via TickAllObjects from the render thread can end up blocking the render thread if not careful.

Added a particle effect that triggers when a unit takes damage. When their shields are hit a silver particle effect occurs and when they take actual damage a red particle effect occurs.

Added an ObjectSpritesheet that will contain sprites for general use objects such as an arrow for use in the Shoot ability's visual effect.